move convo mappers to separate file
